feat: App.tsx 라우팅 분리 및 DashboardPage 컴포넌트 추출
