+++ fe/package.json
@@ -21,6 +21,7 @@
     "lucide-react": "^1.7.0",
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
+    "react-router-dom": "^7.14.0",
     "recharts": "^3.8.1",
     "shadcn": "^4.2.0",
     "tailwind-merge": "^3.5.0",
